WIP Rename (...).cobalt.lib.fmt package to (...).cobalt.lib.doubles Pt1
