+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/Assert.java
