+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/BigNumDtoa.java
