+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/Bignum.java
