+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/DiyFp.java
