+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/DoubleToStringConverter.java
