+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/FastDtoa.java
