+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/FixedDtoa.java
