+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/PowersOfTenCache.java
