+++ src/doubles/java/org/squiddev/cobalt/lib/doubles/UnsignedInt.java
