+++ src/test/java/org/squiddev/cobalt/lib/doubles/BignumTest.java
