+++ src/test/java/org/squiddev/cobalt/lib/doubles/DiyFpTest.java
