+++ src/test/java/org/squiddev/cobalt/lib/doubles/DoubleConversionCompatibility.java
